+++ compiler/src/iree/compiler/Translation/CMakeLists.txt
@@ -10,43 +10,4 @@
 
 iree_add_all_subdirs()
 
-iree_cc_library(
-  NAME
-    HALExecutable
-  HDRS
-    "HALExecutable.h"
-  SRCS
-    "HALExecutable.cpp"
-  DEPS
-    LLVMSupport
-    MLIRIR
-    MLIRPass
-    MLIRSupport
-    MLIRTranslateLib
-    iree::compiler::Dialect::HAL::Target
-    iree::compiler::Dialect::HAL::Transforms
-    iree::compiler::Utils
-  PUBLIC
-)
-
-iree_cc_library(
-  NAME
-    IREEVM
-  HDRS
-    "IREEVM.h"
-  SRCS
-    "IREEVM.cpp"
-  DEPS
-    MLIRPass
-    MLIRTranslateLib
-    iree::compiler::ConstEval
-    iree::compiler::Dialect::Flow::IR
-    iree::compiler::Dialect::HAL::Target
-    iree::compiler::Dialect::VM::Target::C
-    iree::compiler::Pipelines
-    iree::compiler::Pipelines::Options
-    iree::compiler::Utils
-  PUBLIC
-)
-
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
