--- samples/custom_modules/CMakeLists.txt
@@ -1,70 +1,0 @@
-################################################################################
-# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# samples/custom_modules/BUILD                                                 #
-#                                                                              #
-# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
-# CMake-only content.                                                          #
-#                                                                              #
-# To disable autogeneration for this file entirely, delete this header.        #
-################################################################################
-
-if(NOT IREE_TARGET_BACKEND_VMVX OR
-   NOT IREE_HAL_DRIVER_VMVX)
-  return()
-endif()
-
-iree_add_all_subdirs()
-
-iree_cc_library(
-  NAME
-    module
-  HDRS
-    "module.h"
-  SRCS
-    "module.cc"
-  DEPS
-    iree::base
-    iree::base::cc
-    iree::hal
-    iree::modules::hal
-    iree::vm
-    iree::vm::cc
-  PUBLIC
-)
-
-iree_bytecode_module(
-  NAME
-    custom_modules_test_module
-  SRC
-    "custom_modules_test.mlir"
-  C_IDENTIFIER
-    "iree_samples_custom_modules_test_module"
-  TRANSLATE_TOOL
-    iree_samples_custom_modules_dialect_custom-translate
-  FLAGS
-    "--iree-mlir-to-vm-bytecode-module"
-    "--iree-hal-target-backends=vmvx"
-  PUBLIC
-)
-
-iree_cc_test(
-  NAME
-    custom_modules_test
-  SRCS
-    "custom_modules_test.cc"
-  DEPS
-    ::custom_modules_test_module_c
-    ::module
-    iree::base
-    iree::base::logging
-    iree::hal
-    iree::hal::vmvx::registration
-    iree::modules::hal
-    iree::testing::gtest
-    iree::testing::gtest_main
-    iree::vm
-    iree::vm::bytecode_module
-    iree::vm::cc
-)
-
-### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
