--- samples/custom_modules/dialect/CMakeLists.txt
@@ -1,115 +1,0 @@
-################################################################################
-# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# samples/custom_modules/dialect/BUILD                                         #
-#                                                                              #
-# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
-# CMake-only content.                                                          #
-#                                                                              #
-# To disable autogeneration for this file entirely, delete this header.        #
-################################################################################
-
-iree_add_all_subdirs()
-
-iree_cc_library(
-  NAME
-    dialect
-  HDRS
-    "conversion_patterns.h"
-    "custom_dialect.h"
-    "custom_ops.h.inc"
-  SRCS
-    "conversion_patterns.cc"
-    "custom_dialect.cc"
-    "custom_ops.cc.inc"
-  DEPS
-    ::custom_imports
-    ::custom_ops_gen
-    LLVMSupport
-    MLIRIR
-    MLIRParser
-    MLIRPass
-    MLIRSideEffectInterfaces
-    MLIRTransforms
-    iree::compiler::Dialect::HAL::Conversion
-    iree::compiler::Dialect::HAL::IR
-    iree::compiler::Dialect::Util::IR
-    iree::compiler::Dialect::VM::Conversion
-  PUBLIC
-)
-
-iree_tablegen_library(
-  NAME
-    custom_ops_gen
-  TD_FILE
-    "custom_ops.td"
-  OUTS
-    --gen-op-decls custom_ops.h.inc
-    --gen-op-defs custom_ops.cc.inc
-)
-
-iree_c_embed_data(
-  NAME
-    custom_imports
-  SRCS
-    "custom.imports.mlir"
-  C_FILE_OUTPUT
-    "custom.imports.c"
-  H_FILE_OUTPUT
-    "custom.imports.h"
-  IDENTIFIER
-    "iree_custom_imports"
-  FLATTEN
-  PUBLIC
-)
-
-iree_cc_library(
-  NAME
-    init_dialect
-  HDRS
-    "init_dialect.h"
-  DEPS
-    ::dialect
-    MLIRIR
-  PUBLIC
-)
-
-iree_cc_binary(
-  NAME
-    custom-opt
-  SRCS
-    "custom-opt-main.cc"
-  DEPS
-    ::init_dialect
-    LLVMSupport
-    MLIRIR
-    MLIROptLib
-    MLIRSupport
-    iree::compiler::Tools::init_passes_and_dialects
-    iree::compiler::Tools::init_targets
-)
-
-iree_cc_binary(
-  NAME
-    custom-translate
-  SRCS
-    "custom-translate-main.cc"
-  DEPS
-    ::init_dialect
-    LLVMSupport
-    MLIRIR
-    MLIRPass
-    MLIRSCFTransforms
-    MLIRSupport
-    MLIRTranslateLib
-    iree::compiler::Dialect::VM::Target::Bytecode
-    iree::compiler::Dialect::VM::Target::init_targets
-    iree::compiler::Tools::init_compiler_modules
-    iree::compiler::Tools::init_iree_passes_and_dialects
-    iree::compiler::Tools::init_mlir_passes_and_dialects
-    iree::compiler::Tools::init_targets
-    iree::compiler::Tools::init_translations
-    iree::compiler::Tools::init_xla_dialects
-    iree::compiler::Translation::IREEVM
-)
-
-### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
