--- samples/custom_modules/dialect/test/BUILD
@@ -1,30 +1,0 @@
-# Copyright 2019 The IREE Authors
-#
-# Licensed under the Apache License v2.0 with LLVM Exceptions.
-# See https://llvm.org/LICENSE.txt for license information.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-load("//build_tools/bazel:iree_lit_test.bzl", "iree_lit_test_suite")
-load("//build_tools/bazel:enforce_glob.bzl", "enforce_glob")
-
-package(
-    default_visibility = ["//visibility:public"],
-    features = ["layering_check"],
-    licenses = ["notice"],  # Apache 2.0
-)
-
-iree_lit_test_suite(
-    name = "lit",
-    srcs = enforce_glob(
-        [
-            "conversion.mlir",
-            "custom_ops.mlir",
-        ],
-        include = ["*.mlir"],
-    ),
-    cfg = "//samples:lit.cfg.py",
-    tools = [
-        "//samples/custom_modules/dialect:custom-opt",
-        "@llvm-project//llvm:FileCheck",
-    ],
-)
